Move content of ImportData.test into GeantImporter.test
